requirements.txt
aws-cdk-lib==2.142.1
aws-cdk.aws-msk-alpha==2.142.1a0
constructs>=10.0.0,<11.0.0
